backend/requirements.txt
# FastAPI app
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
requests
pillow

# LangChain core
langchain
langchain-openai
langchain-anthropic
langchain-community

# LangChain tools/utilities
duckduckgo-search
wikipedia

# Used by LangChain's WikipediaAPIWrapper
beautifulsoup4


#  wikipedia
# fastapi
# uvicorn
# fastapi
# uvicorn
# pydantic
# python-dotenv
# langchain
# langchain-openai
# langchain-anthropic
# langchain-community
# google-generativeai
# langchain-google-genai

